requirements.txt
groq>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
jsonlines>=3.0.0
